refactor: share audio source representation with mix rendering
